+++ test/unit/property-list.test.js
@@ -499,4 +499,106 @@ describe('PropertyList', function () {
             expect(PropertyList.isPropertyList()).to.be(false);
         });
     });
+
+    describe('toObject', function () {
+        var FakeType = function (options) {
+            this.keyAttr = options.keyAttr;
+            this.value = options.value;
+            this.disabled = options.disabled;
+        };
+
+        FakeType._postman_propertyIndexKey = 'keyAttr';
+        FakeType._postman_propertyIndexCaseInsensitive = true;
+        FakeType._postman_propertyAllowsMultipleValues = true;
+        FakeType.prototype.valueOf = function () {
+            return this.value;
+        };
+
+        it('should return a pojo', function () {
+            var list = new PropertyList(FakeType, {}, [{
+                keyAttr: 'key1',
+                value: 'val1'
+            }, {
+                keyAttr: 'key1',
+                value: 'val2'
+            }, {
+                keyAttr: 'key2',
+                value: 'val3'
+            }]);
+
+            expect(list.toObject()).to.eql({
+                key1: 'val2',
+                key2: 'val3'
+            });
+        });
+
+        it('should be able to override case sensitivity return a pojo', function () {
+            var list = new PropertyList(FakeType, {}, [{
+                keyAttr: 'key1',
+                value: 'val1'
+            }, {
+                keyAttr: 'KEY1',
+                value: 'val2',
+                disabled: true
+            }, {
+                keyAttr: 'key2',
+                value: 'val3'
+            }]);
+
+            expect(list.toObject(null, true)).to.eql({
+                key1: 'val1',
+                KEY1: 'val2',
+                key2: 'val3'
+            });
+        });
+
+        it('should be able to exclude disabled items return a pojo', function () {
+            var list = new PropertyList(FakeType, {}, [{
+                keyAttr: 'key1',
+                value: 'val1'
+            }, {
+                keyAttr: 'key2',
+                value: 'val2',
+                disabled: true
+            }, {
+                keyAttr: 'key3',
+                value: 'val3'
+            }]);
+
+            expect(list.toObject(true)).to.eql({
+                key1: 'val1',
+                key3: 'val3'
+            });
+        });
+    });
+
+    describe('.get', function () {
+        var FakeType = function (options) {
+            this.keyAttr = options.keyAttr;
+            this.value = options.value;
+            this.disabled = options.disabled;
+        };
+
+        FakeType._postman_propertyIndexKey = 'keyAttr';
+        FakeType._postman_propertyIndexCaseInsensitive = true;
+        FakeType._postman_propertyAllowsMultipleValues = false;
+        FakeType.prototype.valueOf = function () {
+            return this.value;
+        };
+
+        it('should return a pojo', function () {
+            var list = new PropertyList(FakeType, {}, [{
+                keyAttr: 'key1',
+                value: 'val1'
+            }, {
+                keyAttr: 'key1',
+                value: 'val2'
+            }, {
+                keyAttr: 'key2',
+                value: 'val3'
+            }]);
+
+            expect(list.get('key1')).to.eql('val2');
+        });
+    });
 });