+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * Added `PropertyBase.findInParents` which finds a given property in the closest parent.
 * Added `PropertyBase.findParentContaining` which finds the closest parent containing the given property.
 * Added `RequestAuth.current()` which returns an instance of the currently selected auth.
+* Added `.toObject` and `.toString` method to PropertyList to work for Types that has a .valueOf and `.unparse` defined.
+* Added HeaderList and **discontinued** `Header.headerValue` in favour of `HeaderList.get`
 
 #### 1.1.0 (April 03, 2017)
 * Enhanced the `PropertyList` to allow keys with multiple values